Range_Finder Example
Put this folder where you put your Arduino projects. The Lidar finder pinout#3=TX is currently connected to pin 17=RX Elegoo and pinout#4=RX is connected to 18=TX of the Elegoo Mega2560, and 5 volts and ground to their respective spots.

This will print out the following information to COM when it is loaded to your Elegoo. 
https://i.imgur.com/F5aoDNx.jpeg