i2c Xodball - in progress
This is a XOD file in progress. I am not good with programming so I thought a graphical tool would work better. It took me a few days but I finally got the hardware working.

I think it writes the value 0 to the i2c and then it ends its transmission to it and starts reading from it. It will request the first 5 bytes. I am not sure what information I'm trying to find as I'm not understanding this part yet but the information I need to get this working I believe is contained here.
https://files.seeedstudio.com/wiki/Grove-TF_Mini_LiDAR/res/SJ-PM-TF-Luna-A03-Product-Manual.pdf